Added a HEAD request to check the Content-Type
In the accept method we now execute a HEAD request to determine the
Content-Type of the file. We accept this download only, if the
Content-Type starts with "video/". This way we can make sure, that we
download an actual video file and not some kind of playlist like the
ones used for DASH or HLS.